Angular CLI - building the application: ( https://github.com/angular/angular-cli/wiki/build )
- 'ng build':
builds a development build (= 'ng build --dev') which means AOT is disabled, hashing only media-files, source-maps will be emitted, css will be put into the *.js-code, every junk will get its own name, build-optimization will not be triggered

- 'ng build --dev --env=prod':
('env' is deprecated !!) uses the development environment (disabled AOT, hashing only media-files etc.) and 'environment.prod.ts' will be used with its settings (= 'ng build --target=development --env=prod')

- 'ng build --prod':
builds the production files to be copied to the server. This uses AOT-compilation, the settings in/with 'environment.prod.ts' are activated, every output-element gets a hash (used for caching...), the source-maps are deactivated and will not be emitted to get a slim output, css will be written in a separate file (only for the 'styless.scss' in root directory!), names for junks will be deactivated (if there are lazy-loaded modules, the junk-names will get deactivated, so no hacker will get information, which junk belongs to which route) and the build optimizer will go through the output, which means the output will be a lot of smaller in size.

- 'ng build --base-href=/subfolder_of_www-root/':
the application must not be saved in the root of a webdomain, it can easily be deployed in subfolders

- 'ng build --watch':
the application will be build. If any file changes, it will build again (but very much faster than a manual triggered build again)