At the moment a 'user-name.component' is a child of 'user-list.component'. The 'user-name.component' takes a value for the attribute 'name' through binding with the decorator '@Input' from outside - here: from 'user-list.component'. These values can be modified - here with 'chgName ()' which has the decorator @HostListener on 'click'.
Additional, there is a decorator '@HostBinding ()' on the css-class 'user-name': if 'isAdminUser = false', the css-class will be bound to the host (the 'user-name.component'). Test it with Augury, change "Benutzer Eins" to "Jochen Gebsattel". ATTENTION: seems not to work like before Angular 6, internal value-changes (with the '@HostListener') do not trigger the '@Input'! '@Input' is obviously triggered only with external input.

So the parent-component tells the child-component, which value will be initially in 'name'.

With '@Output', an EventEmitter can be declared and used to send information to the parent, here with 'this.nameChanged.emit( this.name )'.